+++ pytorch_lightning/loops/dataloader/evaluation_loop.py
@@ -243,39 +243,27 @@ class EvaluationLoop(DataLoaderLoop):
         assert self._results is not None
         self._results.to(device=self.trainer.lightning_module.device)
 
-        if self.trainer.testing:
-            self.trainer._call_callback_hooks("on_test_start", *args, **kwargs)
-            self.trainer._call_lightning_module_hook("on_test_start", *args, **kwargs)
-            self.trainer._call_strategy_hook("on_test_start", *args, **kwargs)
-        else:
-            self.trainer._call_callback_hooks("on_validation_start", *args, **kwargs)
-            self.trainer._call_lightning_module_hook("on_validation_start", *args, **kwargs)
-            self.trainer._call_strategy_hook("on_validation_start", *args, **kwargs)
+        hook_name = "on_test_start" if self.trainer.testing else "on_validation_start"
+        self.trainer._call_callback_hooks(hook_name, *args, **kwargs)
+        self.trainer._call_lightning_module_hook(hook_name, *args, **kwargs)
+        self.trainer._call_strategy_hook(hook_name, *args, **kwargs)
 
     def _on_evaluation_model_eval(self) -> None:
         """Sets model to eval mode."""
-        if self.trainer.testing:
-            self.trainer._call_lightning_module_hook("on_test_model_eval")
-        else:
-            self.trainer._call_lightning_module_hook("on_validation_model_eval")
+        hook_name = "on_test_model_eval" if self.trainer.testing else "on_validation_model_eval"
+        self.trainer._call_lightning_module_hook(hook_name)
 
     def _on_evaluation_model_train(self) -> None:
         """Sets model to train mode."""
-        if self.trainer.testing:
-            self.trainer._call_lightning_module_hook("on_test_model_train")
-        else:
-            self.trainer._call_lightning_module_hook("on_validation_model_train")
+        hook_name = "on_test_model_train" if self.trainer.testing else "on_validation_model_train"
+        self.trainer._call_lightning_module_hook(hook_name)
 
     def _on_evaluation_end(self, *args: Any, **kwargs: Any) -> None:
         """Runs ``on_{validation/test}_end`` hook."""
-        if self.trainer.testing:
-            self.trainer._call_callback_hooks("on_test_end", *args, **kwargs)
-            self.trainer._call_lightning_module_hook("on_test_end", *args, **kwargs)
-            self.trainer._call_strategy_hook("on_test_end", *args, **kwargs)
-        else:
-            self.trainer._call_callback_hooks("on_validation_end", *args, **kwargs)
-            self.trainer._call_lightning_module_hook("on_validation_end", *args, **kwargs)
-            self.trainer._call_strategy_hook("on_validation_end", *args, **kwargs)
+        hook_name = "on_test_end" if self.trainer.testing else "on_validation_end"
+        self.trainer._call_callback_hooks(hook_name, *args, **kwargs)
+        self.trainer._call_lightning_module_hook(hook_name, *args, **kwargs)
+        self.trainer._call_strategy_hook(hook_name, *args, **kwargs)
 
         # reset the logger connector state
         self.trainer._logger_connector.reset_results()
@@ -286,12 +274,9 @@ class EvaluationLoop(DataLoaderLoop):
         self.trainer._call_callback_hooks("on_epoch_start", *args, **kwargs)
         self.trainer._call_lightning_module_hook("on_epoch_start", *args, **kwargs)
 
-        if self.trainer.testing:
-            self.trainer._call_callback_hooks("on_test_epoch_start", *args, **kwargs)
-            self.trainer._call_lightning_module_hook("on_test_epoch_start", *args, **kwargs)
-        else:
-            self.trainer._call_callback_hooks("on_validation_epoch_start", *args, **kwargs)
-            self.trainer._call_lightning_module_hook("on_validation_epoch_start", *args, **kwargs)
+        hook_name = "on_test_epoch_start" if self.trainer.testing else "on_validation_epoch_start"
+        self.trainer._call_callback_hooks(hook_name, *args, **kwargs)
+        self.trainer._call_lightning_module_hook(hook_name, *args, **kwargs)
 
     def _evaluation_epoch_end(self, outputs: List[EPOCH_OUTPUT]) -> None:
         """Runs ``{validation/test}_epoch_end``"""
@@ -303,10 +288,8 @@ class EvaluationLoop(DataLoaderLoop):
         )
 
         # call the model epoch end
-        if self.trainer.testing:
-            self.trainer._call_lightning_module_hook("test_epoch_end", output_or_outputs)
-        else:
-            self.trainer._call_lightning_module_hook("validation_epoch_end", output_or_outputs)
+        hook_name = "test_epoch_end" if self.trainer.testing else "validation_epoch_end"
+        self.trainer._call_lightning_module_hook(hook_name, output_or_outputs)
 
     def _on_evaluation_epoch_end(self) -> None:
         """Runs ``on_{validation/test}_epoch_end`` hook."""